Move pkg/subctl/operator/submarinerop to pkg/operator
