+++ pkg/operator/deployment/ensure.go
