+++ pkg/operator/scc/ensure.go
